backend/scripts/requirements.txt
yookassa
python-dotenv